+++ CNCSwebApiProject/Dto/UserAccountsDtos/UserAccountChangePasswordDto.cs
@@ -1,11 +1,10 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace CNCSwebApiProject.Dto;
+namespace CNCSwebApiProject.Dto.UserAccountsDtos;
 
 public class UserAccountChangePasswordDto
 {
-public int Id { get; set; } 
 [Required]
 public required string NewPassword { get; set; }
