+++ CNCSwebApiProject/Services/UserAccountService/IUserAccountService.cs
@@ -1,5 +1,5 @@
 using System;
-using CNCSwebApiProject.Dto;
+using CNCSwebApiProject.Dto.UserAccountsDtos;
 
 namespace CNCSwebApiProject.Services.UserAccountService;
 
@@ -8,8 +8,8 @@ public interface IUserAccountService
     Task<IEnumerable<UserAccountListDto>> GetAllAsync(); //get all user account
     Task<UserAccountGetAndUpdateDto?> GetAsync(int id); //get single user account
     Task<bool> AddAsync(UserAccountCreateDto userAccount); // Add new user account
-    Task<bool> UpdateDetailsAsync(UserAccountGetAndUpdateDto userAccount); // Update existing user account
-    Task<bool> UpdatePasswordAsync(UserAccountChangePasswordDto userAccountChangePasswordDto);
+    Task<bool> UpdateDetailsAsync(int UserAccount_Id, UserAccountGetAndUpdateDto userAccount); // Update existing user account
+    Task<bool> UpdatePasswordAsync(int UserAccount_Id, UserAccountChangePasswordDto userAccountChangePasswordDto);
     Task<bool> DeleteAsync(int id); // Delete user account by ID
     Task<bool> IsUserExistsAsync(string Username, int id);
 }
